\name{californiacatalog}
\docType{data}
\alias{californiacatalog}
\title{Sample catalog of North California earthquakes}
\description{
Sample catalog of North California earthquakes of magnitude at least 3.0 from year 1968 to year 2012.}
\usage{californiacatalog}
\format{a data matrix with 18,545 observations and 5 variables  \code{time}, \code{lat}, \code{long}, \code{z}, \code{magn1}.
}
\examples{
data(californiacatalog)
str(californiacatalog)
}
\source{Northern California Earthquake Data Center.}
\references{Northern California Earthquake Catalog Search: \url{http://www.ncedc.org/ncedc/catalog-search.html}.}
\keyword{datasets}
\keyword{earthquake}
