% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily.etasclass.R
\name{daily.etasclass}
\alias{daily.etasclass}
\title{Daily estimation on a space grid}
\usage{
daily.etasclass(x, ngrid = 201, nclass = 20, tfixed = 0, flag.log = FALSE, ...)
}
\arguments{
\item{x}{an etasclass object}

\item{ngrid}{subdivisions of x and y axis for grid computation of intensities}

\item{nclass}{number of class for horizontal and vertical axes of the output grid}

\item{tfixed}{day of computation}

\item{flag.log}{if log intensity must be used}

\item{...}{other optional parameters}
}
\value{
a grid with daily theoretical intensities
}
\description{
A daily estimation on a space grid is made
}
