\name{italycatalog}
\docType{data}
\alias{italycatalog}
\title{Small sample catalog of italian earthquakes}
\description{
A small sample catalog of italian earthquakes.}
\usage{italycatalog}
\format{a data matrix with 2158 observations and 5 variables \code{time}, \code{lat}, \code{long}, \code{z}, \code{magn1}.
}
\examples{
data(italycatalog)
str(italycatalog)
}
\source{INGV ISIDE Data Base.}
\references{INGV home page \url{http://www.ingv.it}}
\keyword{datasets}
\keyword{earthquake}
