% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeupdate.etasclass.R
\name{timeupdate.etasclass}
\alias{timeupdate.etasclass}
\title{A time updating of an etasclass objects}
\usage{
timeupdate.etasclass(object, params.estimation = FALSE, ...)
}
\arguments{
\item{object}{an etasclass object  obtained from \code{etasFLP} versions 2.2 or newer that will be updated for a new time window with new events.}

\item{params.estimation}{logical. if \code{TRUE} parameters will be  estimated again even if quickly, with few optimizations steps. Elsewhere ETAS estimates of input object will be mantained}

\item{...}{optional arguments that will override the corresponding arguments in \code{object}, possibly including a new catalog input or a new \code{tmax}}
}
\value{
a new \code{etasclass} object
}
\description{
New in version 2.2. A time updating of an etasclass objects: a very experimental version that can be used only on etasclass objects obtained from etasflp versions 2.2 or newer.
}
\details{
It is a beta version. 
A new ETAS model is fitted to a previous object of class etasclass with a new catalog which must be a catalog which extends the previous one on a wider time window, that is a catalog with new observations.  

As a default a new quick execution is made, with one quick iteration for parameter updating and an iteration for background density estimation.
}
\seealso{
\code{\link{update.etasclass}}
}
